requirements.txt
# flask==3.0.0
# Flask-Cors==4.0.0
pandas==2.2.0
# requests==2.31.0
# pyarrow==15.0.0
# openpyxl==3.1.2
langchain==0.1.6
openai==1.12.0
# Werkzeug==3.0.1
langchain-experimental==0.0.50
# tabulate==0.9.0

python-dotenv == 1.0.1
langchain-community==0.0.19
langchain_openai == 0.0.5